Move factorybot around to see if that helps - ADD 'end' TO FACTORY
