+++ spec/factories/users.rb
@@ -3,3 +3,4 @@ FactoryBot.define do
     email { Faker::Internet.email }
     password  { Faker::Internet.password }
   end
+end
